@@ -18,55 +18,50 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
+  animated_splash_screen: ^1.1.0
+  camera: ^0.9.4
+  cupertino_icons: ^1.0.4
+  double_back_to_close_app: ^2.0.1
+  filesystem_picker: ^2.0.0
   flutter:
     sdk: flutter
-  animated_splash_screen: ^1.0.1+2
-  camera: ^0.5.8+5
-  cupertino_icons: ^0.1.3
-  double_back_to_close_app: ^2.0.1
-  esys_flutter_share: ^1.0.2
-  ext_storage: ^1.0.3
-  flutter_full_pdf_viewer: ^1.0.6
-  flutter_share: ^1.0.5
-  flutter_svg: ^0.19.3
-  fluttertoast: ^8.0.3
-  folder_picker: ^0.3.0
-  gallery_saver: null
-  hive: ^1.4.4+1
-  hive_flutter: ^0.3.1
-  image_cropper: ^1.3.1
-  image_picker: ^0.6.7+21
-  intl: ^0.15.7
-  introduction_screen: ^1.0.9
-  open_file: ^2.0.3
-  package_info: ^2.0.0
-  path: ^1.7.0
-  path_provider: ^1.6.27
-  pdf: ^1.9.0
-  pdf_viewer_plugin: ^1.1.1
-  permission: ^0.1.7
-  permission_handler: null
-  photofilters: ^2.0.1
-  provider: ^4.3.3
-  quick_actions: ^0.4.0+10
-  rating_dialog: ^1.0.0
-  share: ^0.6.5+4
-  shared_preferences: ^0.5.12+4
-  source_span: ^1.7.0
-  storage_path: ^0.2.0
-  url_launcher: ^5.7.10
-  wiredash: ^0.5.0
-  flutter_launcher_icons: ^0.7.2+1
-  flutter_spinkit: ^5.0.0
+  # flutter_full_pdf_viewer: ^1.0.6
+  flutter_launcher_icons: ^0.9.2
+  flutter_spinkit: ^5.1.0
+  flutter_svg: ^0.23.0+1
+  fluttertoast: ^8.0.8
+  gallery_saver: ^2.3.2
+  hive: ^2.0.4
+  hive_flutter: ^1.1.0
+  image_cropper: ^1.4.1
+  image_picker: ^0.8.4+4
+  intl: ^0.17.0
+  introduction_screen: ^2.1.0
+  open_file: ^3.2.1
+  package_info: ^2.0.2
+  path: ^1.8.0
+  path_provider: ^2.0.7
+  pdf: ^3.6.2
+  pdf_viewer_plugin: ^2.0.1
+  permission_handler: ^8.3.0
+  photofilters: ^3.0.1
+  provider: ^6.0.1
+  quick_actions: ^0.6.0+7
+  rating_dialog: ^2.0.3
+  share: ^2.0.4
+  shared_preferences: ^2.0.8
+  source_span: ^1.8.1
+  url_launcher: ^6.0.13
+  wiredash: ^0.7.0+1
 dev_dependencies:
+  build_runner: ^2.1.5
   flutter_test:
     sdk: flutter
-  build_runner: null
-  hive_generator: ^0.8.2
-  lint: ^1.0.0
+  hive_generator: ^1.1.1
+lint: ^1.7.2
 
 flutter_icons:
   android: true
@@ -115,4 +110,4 @@ flutter:
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
 dependencies_overrides:
-  collection: ^1.14.12
+  collection: ^1.15.0
