@@ -1,12 +1,12 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_full_pdf_viewer/flutter_full_pdf_viewer.dart';
+// import 'package:flutter_full_pdf_viewer/flutter_full_pdf_viewer.dart';
 
 class PdfPreviewScreen extends StatefulWidget {
   final String path;
 
-  const PdfPreviewScreen({this.path});
+  const PdfPreviewScreen({required this.path});
 
   @override
   _PdfPreviewScreenState createState() => _PdfPreviewScreenState();
@@ -27,11 +27,12 @@ class _PdfPreviewScreenState extends State<PdfPreviewScreen> {
 
   @override
   Widget build(BuildContext context) {
-    return PDFViewerScaffold(
-      path: widget.path,
-      // appBar: AppBar(
-      //   backgroundColor: Colors.blue,
-      // ),
-    );
+    return Container();
+    // PDFViewerScaffold(
+    // path: widget.path,
+    // appBar: AppBar(
+    //   backgroundColor: Colors.blue,
+    // ),
+    // );
   }
 }
