@@ -7,8 +7,7 @@ class DarkThemeProvider with ChangeNotifier {
 
   bool get darkTheme => _darkTheme;
 
-
-  set darkTheme(bool value) {
+  set darkTheme(bool? value) {
     _darkTheme = value ?? false;
     darkThemePreference.setSharedPreferenceValue("themeMode", value);
     notifyListeners();
