@@ -18,7 +18,7 @@ class _ContactDeveloperScreenState extends State<ContactDeveloperScreen> {
   Future<void> _fetchContributors() async {
     const contributorsAPIUrl =
         'https://api.github.com/repos/smaranjitghose/DocLense/contributors';
-    final response = await http.get(contributorsAPIUrl);
+    final response = await http.get(Uri(host: contributorsAPIUrl));
 
     if (response.statusCode == 200) {
       setState(() {
@@ -34,7 +34,7 @@ class _ContactDeveloperScreenState extends State<ContactDeveloperScreen> {
   }
 
   bool _isLoading = true;
-  
+
   @override
   void initState() {
     super.initState();
@@ -49,161 +49,174 @@ class _ContactDeveloperScreenState extends State<ContactDeveloperScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body:  _isLoading ?const SpinKitRotatingCircle(
-  color: Colors.blue,
-) : SafeArea(
-          child: SingleChildScrollView(
-        child: Container(
-          margin: const EdgeInsets.only(top: 30),
-          child: Column(
-            children: <Widget>[
-              Container(
-                alignment: Alignment.topCenter,
-                child: Row(
-                  mainAxisSize: MainAxisSize.min,
+      body: _isLoading
+          ? const SpinKitRotatingCircle(
+              color: Colors.blue,
+            )
+          : SafeArea(
+              child: SingleChildScrollView(
+              child: Container(
+                margin: const EdgeInsets.only(top: 30),
+                child: Column(
                   children: <Widget>[
-                    const Text(
-                      "TEAM",
-                      style:
-                          TextStyle(fontSize: 25, fontWeight: FontWeight.w700),
+                    Container(
+                      alignment: Alignment.topCenter,
+                      child: Row(
+                        mainAxisSize: MainAxisSize.min,
+                        children: <Widget>[
+                          const Text(
+                            "TEAM",
+                            style: TextStyle(
+                                fontSize: 25, fontWeight: FontWeight.w700),
+                          ),
+                          Container(
+                            margin: const EdgeInsets.only(left: 8),
+                            child: const Text(
+                              "2021",
+                              style: TextStyle(
+                                fontSize: 25,
+                                fontWeight: FontWeight.w700,
+                                // color: Colors.orange[900],
+                              ),
+                            ),
+                          )
+                        ],
+                      ),
                     ),
                     Container(
-                      margin: const EdgeInsets.only(left: 8),
+                      margin: const EdgeInsets.only(top: 15),
+                      alignment: Alignment.topCenter,
                       child: const Text(
-                        "2021",
+                        "Presenting The Team",
                         style: TextStyle(
-                          fontSize: 25,
-                          fontWeight: FontWeight.w700,
-                          // color: Colors.orange[900],
-                        ),
+                            fontSize: 18, fontWeight: FontWeight.w500),
                       ),
-                    )
-                  ],
-                ),
-              ),
-              Container(
-                margin: const EdgeInsets.only(top: 15),
-                alignment: Alignment.topCenter,
-                child: const Text(
-                  "Presenting The Team",
-                  style: TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
-                ),
-              ),
-              _buildTeamTitle('DEVELOPERS'),
-              Container(
-                height: 170,
-                margin: const EdgeInsets.only(top: 20),
-                child: ListView.builder(
-                  shrinkWrap: true,
-                  scrollDirection: Axis.horizontal,
-                  itemCount: devTeam.length,
-                  itemBuilder: (BuildContext context, int index) => Container(
-                    margin: const EdgeInsets.only(left: 15, right: 15),
-                    child: Column(
-                      children: <Widget>[
-                        _buildprofileImage(devTeam[index]["imgPath"]),
-                        Container(
-                          margin: const EdgeInsets.only(top: 10),
+                    ),
+                    _buildTeamTitle('DEVELOPERS'),
+                    Container(
+                      height: 170,
+                      margin: const EdgeInsets.only(top: 20),
+                      child: ListView.builder(
+                        shrinkWrap: true,
+                        scrollDirection: Axis.horizontal,
+                        itemCount: devTeam.length,
+                        itemBuilder: (BuildContext context, int index) =>
+                            Container(
+                          margin: const EdgeInsets.only(left: 15, right: 15),
                           child: Column(
                             children: <Widget>[
-                              Text(
-                                devTeam[index]["name"],
-                              ),
-                              const Padding(padding: EdgeInsets.only(top: 5)),
-                              Row(
-                                children: <Widget>[
-                                  _buildProfileIcon(devTeam[index]["linkedin"],
-                                      'https://img.icons8.com/fluent/48/000000/linkedin-circled.png'),
-                                  const Padding(
-                                      padding: EdgeInsets.only(left: 10)),
-                                  _buildProfileIcon(devTeam[index]["github"],
-                                      'https://img.icons8.com/fluent/50/000000/github.png'),
-                                ],
-                              )
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-              _buildTeamTitle('CONTIBUTORS'),
-              // ignore: prefer_is_empty
-              if (jsonContributors.length > 0)
-                Container(
-                  height: 170,
-                  margin: const EdgeInsets.only(top: 20),
-                  child: ListView.builder(
-                    shrinkWrap: true,
-                    scrollDirection: Axis.horizontal,
-                    itemCount: jsonContributors.length,
-                    itemBuilder: (BuildContext context, int index) =>
-                        (jsonContributors[index]["login"].toString() ==
-                                    "smaranjitghose" ||
-                                jsonContributors[index]["login"].toString() ==
-                                    "anushbhatia")
-                            ? Container()
-                            : Container(
-                                margin:
-                                    const EdgeInsets.only(left: 15, right: 15),
+                              _buildprofileImage(devTeam[index]["imgPath"]!),
+                              Container(
+                                margin: const EdgeInsets.only(top: 10),
                                 child: Column(
                                   children: <Widget>[
-                                    _buildNetworkprofileImage(
-                                        jsonContributors[index]["avatar_url"]
-                                            .toString()),
-                                    Container(
-                                      margin: const EdgeInsets.only(top: 10),
-                                      child: Column(
-                                        children: <Widget>[
-                                          Text(
-                                            jsonContributors[index]["login"]
-                                                .toString(),
-                                          ),
-                                          const Padding(
-                                              padding: EdgeInsets.only(top: 5)),
-                                          Row(
-                                            children: <Widget>[
-                                              if (jsonContributors[index]
-                                                          ["login"]
-                                                      .toString() ==
-                                                  "Saransh-cpp")
-                                                _buildProfileIcon(
-                                                    "https://www.linkedin.com/in/saransh-chopra-3a6ab11bb",
-                                                    'https://img.icons8.com/fluent/48/000000/linkedin-circled.png')
-                                              else if (jsonContributors[index]
-                                                          ["login"]
-                                                      .toString() ==
-                                                  "nicks101")
-                                                _buildProfileIcon(
-                                                    "https://www.linkedin.com/in/nikki-goel-449563159/",
-                                                    'https://img.icons8.com/fluent/48/000000/linkedin-circled.png')
-                                              else
-                                                Container(),
-                                              const Padding(
-                                                  padding: EdgeInsets.only(
-                                                      left: 10)),
-                                              _buildProfileIcon(
-                                                  jsonContributors[index]
-                                                          ["html_url"]
-                                                      .toString(),
-                                                  'https://img.icons8.com/fluent/50/000000/github.png'),
-                                            ],
-                                          )
-                                        ],
-                                      ),
+                                    Text(
+                                      devTeam[index]["name"]!,
                                     ),
+                                    const Padding(
+                                        padding: EdgeInsets.only(top: 5)),
+                                    Row(
+                                      children: <Widget>[
+                                        _buildProfileIcon(
+                                            devTeam[index]["linkedin"]!,
+                                            'https://img.icons8.com/fluent/48/000000/linkedin-circled.png'),
+                                        const Padding(
+                                            padding: EdgeInsets.only(left: 10)),
+                                        _buildProfileIcon(
+                                            devTeam[index]["github"]!,
+                                            'https://img.icons8.com/fluent/50/000000/github.png'),
+                                      ],
+                                    )
                                   ],
                                 ),
                               ),
-                  ),
-                )
-              else
-                const Center(child: CircularProgressIndicator()),
-            ],
-          ),
-        ),
-      )),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                    _buildTeamTitle('CONTIBUTORS'),
+                    // ignore: prefer_is_empty
+                    if (jsonContributors.length > 0)
+                      Container(
+                        height: 170,
+                        margin: const EdgeInsets.only(top: 20),
+                        child: ListView.builder(
+                          shrinkWrap: true,
+                          scrollDirection: Axis.horizontal,
+                          itemCount: jsonContributors.length,
+                          itemBuilder: (BuildContext context, int index) =>
+                              (jsonContributors[index]["login"].toString() ==
+                                          "smaranjitghose" ||
+                                      jsonContributors[index]["login"]
+                                              .toString() ==
+                                          "anushbhatia")
+                                  ? Container()
+                                  : Container(
+                                      margin: const EdgeInsets.only(
+                                          left: 15, right: 15),
+                                      child: Column(
+                                        children: <Widget>[
+                                          _buildNetworkprofileImage(
+                                              jsonContributors[index]
+                                                      ["avatar_url"]
+                                                  .toString()),
+                                          Container(
+                                            margin:
+                                                const EdgeInsets.only(top: 10),
+                                            child: Column(
+                                              children: <Widget>[
+                                                Text(
+                                                  jsonContributors[index]
+                                                          ["login"]
+                                                      .toString(),
+                                                ),
+                                                const Padding(
+                                                    padding: EdgeInsets.only(
+                                                        top: 5)),
+                                                Row(
+                                                  children: <Widget>[
+                                                    if (jsonContributors[index]
+                                                                ["login"]
+                                                            .toString() ==
+                                                        "Saransh-cpp")
+                                                      _buildProfileIcon(
+                                                          "https://www.linkedin.com/in/saransh-chopra-3a6ab11bb",
+                                                          'https://img.icons8.com/fluent/48/000000/linkedin-circled.png')
+                                                    else if (jsonContributors[
+                                                                index]["login"]
+                                                            .toString() ==
+                                                        "nicks101")
+                                                      _buildProfileIcon(
+                                                          "https://www.linkedin.com/in/nikki-goel-449563159/",
+                                                          'https://img.icons8.com/fluent/48/000000/linkedin-circled.png')
+                                                    else
+                                                      Container(),
+                                                    const Padding(
+                                                        padding:
+                                                            EdgeInsets.only(
+                                                                left: 10)),
+                                                    _buildProfileIcon(
+                                                        jsonContributors[index]
+                                                                ["html_url"]
+                                                            .toString(),
+                                                        'https://img.icons8.com/fluent/50/000000/github.png'),
+                                                  ],
+                                                )
+                                              ],
+                                            ),
+                                          ),
+                                        ],
+                                      ),
+                                    ),
+                        ),
+                      )
+                    else
+                      const Center(child: CircularProgressIndicator()),
+                  ],
+                ),
+              ),
+            )),
     );
   }
 
